requirements.txt
behave==1.2.6
selenium==4.23.1
allure-behave==2.13.5
requests==2.32.3
webdriver-manager==4.0.2
